requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
redis==5.0.7
psycopg2-binary==2.9.9
python-dotenv==1.0.1
